requirements.txt
mysql-connector-python
pandas
playwright